mcp-security-elevate25/run-with-google-adk/requirements.txt
google-adk==0.5.0
uv